refactor: extract camera logic to hook
